backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
supabase
python-dotenv
pydantic-ai
pytest
pytest-asyncio
httpx